+++ cmd/web/main.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"log"
+	"net/http"
+)
+
+func main() {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/", home)
+	mux.HandleFunc("/snippet", showSnippet)
+	mux.HandleFunc("/snippet/create", createSnippet)
+
+	log.Println("Запуск сервера на http://127.0.0.1:4000")
+	err := http.ListenAndServe(":4000", mux)
+	log.Fatal(err)
+}
