+++ ui/html/footer.partial.tmpl
@@ -1,0 +1,3 @@
+{{define "footer"}}
+<footer>Сайт создан на <strong>Golang</strong></footer>
+{{end}}
