+++ ui/html/home.page.tmpl
@@ -1,0 +1,8 @@
+{{template "base" .}}
+
+{{define "title"}}Домашняя страницы{{end}}
+
+{{define "main"}}
+<h2>Последние заметки</h2>
+<p>Пока здесь ничего нет!</p>
+{{end}}
